refactor: moved createRequestHandler and createRoutes up to the HttpServer
