+++ packages/http-server-express/src/ExpressHttpServer.ts
@@ -2,21 +2,11 @@
  * © Copyright 2022 HP Development Company, L.P.
  * SPDX-License-Identifier: MIT
  */
-import {
-	ControllerDecoratorMetadata,
-	HttpServerModule,
-	HttpServerModuleOptions,
-	MethodDecoratorMetadata,
-	ParameterDecoratorMetadata,
-	ParameterSource,
-	RequestHandler
-} from '@davinci/http-server';
+import { HttpServerModule, HttpServerModuleOptions, ParameterSource, RequestHandler } from '@davinci/http-server';
 import express, { Express, Request, Response } from 'express';
 import http, { Server as HttpServer } from 'http';
 import https, { Server as HttpsServer } from 'https';
-import pathUtils from 'path';
 import type { App } from '@davinci/core';
-import { ClassType, MethodReflection } from '@davinci/reflector';
 
 type Server = HttpServer | HttpsServer;
 
@@ -34,7 +24,7 @@ export class ExpressHttpServer extends HttpServerModule<Request, Response, Serve
 
 	async onInit(app) {
 		this.app = app;
-		this.createRoutes();
+		super.createRoutes();
 		this.initHttpServer();
 		return super.getHttpServer().listen(super.moduleOptions?.port ?? 3000);
 	}
@@ -56,52 +46,6 @@ export class ExpressHttpServer extends HttpServerModule<Request, Response, Serve
 			: http.createServer(this.getInstance());
 
 		super.setHttpServer(server);
-	}
-
-	createRoutes() {
-		const controllersReflection = this.app
-			.getControllersWithReflection()
-			.filter(
-				({ reflection }) =>
-					reflection.decorators.some(d => d.module === 'http-server') ||
-					reflection.methods.some(m => m.decorators.some(d => d.module === 'http-server'))
-			);
-
-		return controllersReflection.map(({ Controller, reflection: controllerReflection }) => {
-			const controllerDecoratorMetadata: ControllerDecoratorMetadata = controllerReflection.decorators.find(
-				d => d.module === 'http-server' && d.type === 'controller'
-			);
-			const basePath =
-				controllerDecoratorMetadata?.options?.basePath ?? controllerDecoratorMetadata?.options?.basepath ?? '/';
-
-			let ctrlInstance: typeof Controller;
-			const getControllerInstance = () => {
-				if (ctrlInstance) return ctrlInstance;
-
-				ctrlInstance = new Controller();
-
-				return ctrlInstance;
-			};
-
-			return controllerReflection.methods.map(methodReflection => {
-				const methodDecoratorMetadata: MethodDecoratorMetadata = methodReflection.decorators.find(
-					d => d.module === 'http-server' && d.type === 'route'
-				);
-				const methodName = methodReflection.name;
-
-				if (!methodDecoratorMetadata) return null;
-
-				const {
-					verb,
-					options: { path }
-				} = methodDecoratorMetadata;
-
-				const controller = getControllerInstance();
-				const fullPath = pathUtils.join(basePath, path);
-
-				return this[verb](fullPath, this.createRequestHandler(controller, methodName, methodReflection));
-			});
-		});
 	}
 
 	public reply(response: Response, body: unknown, statusCode?: number) {
@@ -206,35 +150,6 @@ export class ExpressHttpServer extends HttpServerModule<Request, Response, Serve
 
 	public getRequestUrl(request: Request): string {
 		return request.originalUrl;
-	}
-
-	createRequestHandler(controller: InstanceType<ClassType>, methodName: string, methodReflection: MethodReflection) {
-		return async (req: Request, res: Response) => {
-			const parameters = methodReflection.parameters.map(parameterReflection => {
-				const parameterDecoratorMetadata: ParameterDecoratorMetadata = parameterReflection.decorators.find(
-					d => d.module === 'http-server' && d.type === 'parameter'
-				);
-
-				if (parameterDecoratorMetadata) {
-					const { options } = parameterDecoratorMetadata;
-					return this.getRequestParameter({
-						source: options.in,
-						name: options.name ?? parameterReflection.name,
-						request: req
-					});
-				}
-
-				return undefined;
-			});
-
-			try {
-				const result = await controller[methodName](...parameters);
-
-				return this.reply(res, result);
-			} catch (err) {
-				return this.reply(res, { error: true, message: err.message }, 500);
-			}
-		};
 	}
 
 	getRequestParameter({ source, name, request }: { source: ParameterSource; name?: string; request: Request }) {
