+++ packages/http-server/src/HttpServerModule.ts
@@ -2,11 +2,15 @@
  * © Copyright 2022 HP Development Compunknown, L.P.
  * SPDX-License-Identifier: MIT
  */
-import { Module } from '@davinci/core';
-import { RequestHandler, HttpServerModuleOptions, ParameterSource /* CorsOptions */ } from './types';
+import { App, Module } from '@davinci/core';
+import pathUtils from 'path';
+import { ClassType, MethodReflection } from '@davinci/reflector';
+import { HttpServerModuleOptions, ParameterSource, RequestHandler } from './types';
+import { ControllerDecoratorMetadata, MethodDecoratorMetadata, ParameterDecoratorMetadata } from './decorators';
 
 export abstract class HttpServerModule<Request = unknown, Response = unknown, Server = unknown> extends Module {
 	protected httpServer: Server;
+	app: App;
 
 	getModuleId() {
 		return 'http';
@@ -26,6 +30,89 @@ export abstract class HttpServerModule<Request = unknown, Response = unknown, Se
 
 	public setHttpServer(httpServer: Server) {
 		this.httpServer = httpServer;
+	}
+
+	public createRoutes() {
+		const controllersReflection = this.app
+			.getControllersWithReflection()
+			.filter(
+				({ reflection }) =>
+					reflection.decorators.some(d => d.module === 'http-server') ||
+					reflection.methods.some(m => m.decorators.some(d => d.module === 'http-server'))
+			);
+
+		return controllersReflection.map(({ Controller, reflection: controllerReflection }) => {
+			const controllerDecoratorMetadata: ControllerDecoratorMetadata = controllerReflection.decorators.find(
+				d => d.module === 'http-server' && d.type === 'controller'
+			);
+			const basePath =
+				controllerDecoratorMetadata?.options?.basePath ?? controllerDecoratorMetadata?.options?.basepath ?? '/';
+
+			let ctrlInstance: typeof Controller;
+			const getControllerInstance = () => {
+				if (ctrlInstance) return ctrlInstance;
+
+				ctrlInstance = new Controller();
+
+				return ctrlInstance;
+			};
+
+			return controllerReflection.methods.map(methodReflection => {
+				const methodDecoratorMetadata: MethodDecoratorMetadata = methodReflection.decorators.find(
+					d => d.module === 'http-server' && d.type === 'route'
+				);
+				const methodName = methodReflection.name;
+
+				if (!methodDecoratorMetadata) return null;
+
+				const {
+					verb,
+					options: { path }
+				} = methodDecoratorMetadata;
+
+				const controller = getControllerInstance();
+				const fullPath = pathUtils.join(basePath, path);
+
+				return this[verb](fullPath, this.createRequestHandler(controller, methodName, methodReflection));
+			});
+		});
+	}
+
+	public createRequestHandler(
+		controller: InstanceType<ClassType>,
+		methodName: string,
+		methodReflection: MethodReflection
+	) {
+		// eslint-disable-next-line @typescript-eslint/no-this-alias
+		const that = this;
+
+		// using a named function here for better instrumentation and reporting
+		return async function expressRequestHandler(req: Request, res: Response) {
+			const parameters = methodReflection.parameters.map(parameterReflection => {
+				const parameterDecoratorMetadata: ParameterDecoratorMetadata = parameterReflection.decorators.find(
+					d => d.module === 'http-server' && d.type === 'parameter'
+				);
+
+				if (parameterDecoratorMetadata) {
+					const { options } = parameterDecoratorMetadata;
+					return that.getRequestParameter({
+						source: options.in,
+						name: options.name ?? parameterReflection.name,
+						request: req
+					});
+				}
+
+				return undefined;
+			});
+
+			try {
+				const result = await controller[methodName](...parameters);
+
+				return that.reply(res, result);
+			} catch (err) {
+				return that.reply(res, { error: true, message: err.message }, 500);
+			}
+		};
 	}
 
 	// abstract get(handler: Function);
@@ -56,18 +143,31 @@ export abstract class HttpServerModule<Request = unknown, Response = unknown, Se
 	abstract listen(port: string | number, hostname: string, callback?: () => void);
 
 	abstract initHttpServer(): void;
+
 	abstract setInstance(instance: unknown): void;
+
 	abstract getInstance(): void;
+
 	abstract reply(response, body: unknown, statusCode?: number);
+
 	abstract close();
+
 	abstract getRequestHostname(request: Request);
+
 	abstract getRequestMethod(request: Request);
+
 	abstract getRequestUrl(request: Request);
+
 	abstract getRequestParameter(args: { source: ParameterSource; name?: string; request: Request });
+
 	abstract status(response, statusCode: number);
+
 	abstract redirect(response, statusCode: number, url: string);
+
 	abstract setErrorHandler(handler: Function, prefix?: string);
+
 	abstract setNotFoundHandler(handler: Function, prefix?: string);
+
 	abstract setHeader(response, name: string, value: string);
 
 	/* abstract render(response, view: string, options: unknown);
